微博-图文混排

+++ weibo/weibo/Classes/OAuth/Controller/ZSROAuthViewController.m
@@ -66,7 +66,7 @@
     NSRange range = [url rangeOfString:@"code="];
     if (range.length != 0) { // 是回调地址
         // 截取code=后面的参数值
-        int fromIndex = range.location + range.length;
+        NSInteger fromIndex = range.location + range.length;
         NSString *code = [url substringFromIndex:fromIndex];
         
         ZSRLog(@"%@ %@",code,url);
